transport-university-chatbot/backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
langchain==0.1.0
langchain-community==0.0.10
langchain-pinecone==0.0.1

